refactor(intelligence,http): move pr_generation handler to mockforge-intelligence (#555 phase 2)
ADR 0001 marks `pr_generation.rs` as the smallest INTELLIGENCE-bucket
handler (146 LOC) and the cleanest first axum-into-intelligence move:
its only foreign import was `mockforge_intelligence::pr_generation::*`,
which #562 phase 1 had already moved into intelligence.

## What changed

- `git mv crates/mockforge-http/src/handlers/pr_generation.rs →
  crates/mockforge-intelligence/src/handlers/pr_generation.rs`.
- Rewrote the file's `mockforge_intelligence::pr_generation::*` import
  to `crate::pr_generation::*` (now sibling).
- Added `crates/mockforge-intelligence/src/handlers/mod.rs` with a
  `pub mod pr_generation;` declaration and a docstring explaining the
  #555 migration pattern.
- Added `pub mod handlers;` to `mockforge-intelligence/src/lib.rs`.
- `mockforge-intelligence/Cargo.toml` gains `axum` as a workspace dep.
  Cycle-safe: `mockforge-http` already depends on
  `mockforge-intelligence`, never the reverse. This is the first axum
  dep on the intelligence side; subsequent handler moves are now
  unblocked from a Cargo-deps perspective.
- `mockforge-http/src/handlers/pr_generation.rs` is now a one-line
  re-export shim
  (`pub use mockforge_intelligence::handlers::pr_generation::*;`)
  so `mockforge_http::handlers::pr_generation::*` callers — notably
  the router-construction code in `mockforge_http::lib` — keep
  resolving unchanged.

Drive-by: extend `audit.toml` to ignore RUSTSEC-2026-0149 (wasmtime-wasi
`path_open(TRUNCATE)` permission bypass). This is a new advisory in the
same wasmtime-36.x pin family the file already documents as blocked by
the plugin-loader's wasmtime major bump. Pre-commit cargo-audit started
failing on it today because rustsec-advisory-db updated; not introduced
by this PR.

## Verification

- `cargo check -p mockforge-intelligence -p mockforge-http` — clean
- `cargo test -p mockforge-intelligence -p mockforge-http --lib` —
  926 pass (615 http, 311 intelligence)
- `cargo clippy -p mockforge-intelligence -p mockforge-http
  --all-targets -- -D warnings` — clean
- `cargo fmt --all --check` — clean

## Phase context

Per ADR 0001 (docs/adr/0001-mockforge-http-extraction.md), 17
INTELLIGENCE-bucket handlers remain in `mockforge-http/src/handlers/`.
With axum now on the intelligence side, the rest are unblocked from a
Cargo-deps perspective; the per-handler audits required by the ADR
(checking each handler's incidental core/recorder deps) will drive
subsequent #555 phases.
