+++ crates/mockforge-intelligence/Cargo.toml
@@ -22,6 +22,10 @@ mockforge-foundation = { path = "../mockforge-foundation", default-features = fa
 # dep is cycle-safe — it does NOT pull `mockforge-core` back in.
 mockforge-openapi = { path = "../mockforge-openapi" }
 async-trait = { workspace = true }
+# `handlers::*` (HTTP-handler files migrated out of `mockforge-http` under
+# Issue #555) need axum. Cycle-safe: `mockforge-http` already depends on
+# `mockforge-intelligence`, never the reverse.
+axum = { workspace = true }
 base64 = { workspace = true }
 chrono = { workspace = true }
 dirs = "5.0"
